COM-1435 fix from pr comments
