--- src/components/TemplateExample.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { Text } from 'react-native';
-
-interface TemplateExampleProps {
-  template: String,
-}
-
-const TemplateExample = ({ template }: TemplateExampleProps) => (
-  <Text>{template}</Text>
-);
-
-export default TemplateExample;
